+++ game/core/core.php
@@ -27,7 +27,6 @@ require_once "battle_engine.php";
 require_once "raketen.php";
 require_once "graviton.php";
 require_once "debug.php";
-require_once "galaxytool.php";
 require_once "bot.php";
 require_once "botapi.php";
 require_once "coupon.php";
