--- game/galaxytool/.gitignore
@@ -1,1 +1,0 @@
-*.txt
