--- game/loca/de_de/galaxytool.php
@@ -1,22 +1,0 @@
-<?php
-
-// Gebaut Galaxytool
-
-$LOCA["de"]["GALATOOL_TITLE"] = "OGame Information Center";
-$LOCA["de"]["GALATOOL_POINTS"] = "Punkte : #1";
-$LOCA["de"]["GALATOOL_FLEET"] = "Flotte : #1";
-$LOCA["de"]["GALATOOL_RESEARCH"] = "Forschung : #1";
-$LOCA["de"]["GALATOOL_ALLY"] = "Allianz : [#1]";
-$LOCA["de"]["GALATOOL_PLANETS"] = "Planeten";
-$LOCA["de"]["GALATOOL_MOONS"] = "Monds";
-$LOCA["de"]["GALATOOL_DATE"] = "Daten für #1";
-$LOCA["de"]["GALATOOL_GROW"] = "Die Wachstumsraten in den letzten 7 Tagen:";
-$LOCA["de"]["GALATOOL_FALL"] = "Rückgänge in den letzten 7 Tagen:";
-$LOCA["de"]["GALATOOL_INACTIVE"] = "Inaktive oder gesperrt:";
-$LOCA["de"]["GALATOOL_TOTAL"] = "Total : #1";
-$LOCA["de"]["GALATOOL_NOTES"] = "Wenn ein Spieler Anstieg / Rückgang war mehr als 30.000 Punkte, dann wird in Klammern neben dem Namen angezeigt.";
-$LOCA["de"]["GALATOOL_FIND"] = "Finden Sie alle Planeten Spieler : ";
-$LOCA["de"]["GALATOOL_SEARCH"] = "Suchen";
-$LOCA["de"]["GALATOOL_NOT_UPDATED"] = "Noch nicht aktualisiert";
-
-?>
