+++ game/mods/GalaxyTool/loca/de_de/.htaccess
@@ -1,0 +1,1 @@
+Order allow,deny Deny from all
