+++ game/mods/GalaxyTool/loca/en_en/.htaccess
@@ -1,0 +1,1 @@
+Order allow,deny Deny from all
