+++ game/mods/GalaxyTool/loca/en_en/galaxytool.php
@@ -1,0 +1,22 @@
+<?php
+
+// Integrated Galaxytool
+
+$LOCA["en"]["GALATOOL_TITLE"] = "OGame Database";
+$LOCA["en"]["GALATOOL_POINTS"] = "Score : #1";
+$LOCA["en"]["GALATOOL_FLEET"] = "Fleet : #1";
+$LOCA["en"]["GALATOOL_RESEARCH"] = "Research : #1";
+$LOCA["en"]["GALATOOL_ALLY"] = "Ally : [#1]";
+$LOCA["en"]["GALATOOL_PLANETS"] = "Planets";
+$LOCA["en"]["GALATOOL_MOONS"] = "Moons";
+$LOCA["en"]["GALATOOL_DATE"] = "Last updated #1";
+$LOCA["en"]["GALATOOL_GROW"] = "Grow last week:";
+$LOCA["en"]["GALATOOL_FALL"] = "Falls last week:";
+$LOCA["en"]["GALATOOL_INACTIVE"] = "Non-active and banned:";
+$LOCA["en"]["GALATOOL_TOTAL"] = "Total : #1";
+$LOCA["en"]["GALATOOL_NOTES"] = "If a player grow / fall was more than 30,000 points, then it is indicated in parentheses next to the name.";
+$LOCA["en"]["GALATOOL_FIND"] = "Find all player planets : ";
+$LOCA["en"]["GALATOOL_SEARCH"] = "Search";
+$LOCA["en"]["GALATOOL_NOT_UPDATED"] = "Not yet updated";
+
+?>
