+++ game/mods/GalaxyTool/loca/es_es/.htaccess
@@ -1,0 +1,1 @@
+Order allow,deny Deny from all
