+++ game/mods/GalaxyTool/loca/es_es/galaxytool.php
