+++ game/mods/GalaxyTool/loca/fr_fr/.htaccess
@@ -1,0 +1,1 @@
+Order allow,deny Deny from all
