+++ game/mods/GalaxyTool/loca/fr_fr/galaxytool.php
