+++ game/mods/GalaxyTool/loca/it_it/.htaccess
@@ -1,0 +1,1 @@
+Order allow,deny Deny from all
