+++ game/mods/GalaxyTool/loca/it_it/galaxytool.php
