+++ game/mods/GalaxyTool/loca/ru_ru/galaxytool.php
@@ -1,0 +1,23 @@
+<?php
+
+// Встроенная галабаза
+
+$LOCA["ru"]["GALATOOL_TITLE"] = "Информационный Центр Огейм";
+$LOCA["ru"]["GALATOOL_POINTS"] = "Очки : #1";
+$LOCA["ru"]["GALATOOL_FLEET"] = "Флот : #1";
+$LOCA["ru"]["GALATOOL_RESEARCH"] = "Исследования : #1";
+$LOCA["ru"]["GALATOOL_ALLY"] = "Альянс : [#1]";
+$LOCA["ru"]["GALATOOL_PLANETS"] = "Планеты";
+$LOCA["ru"]["GALATOOL_MOONS"] = "Луны";
+$LOCA["ru"]["GALATOOL_DATE"] = "Данные на #1";
+$LOCA["ru"]["GALATOOL_GROW"] = "Приросты за последние 7 дней:";
+$LOCA["ru"]["GALATOOL_FALL"] = "Спады за последние 7 дней:";
+$LOCA["ru"]["GALATOOL_INACTIVE"] = "Неактивные и заблокированные:";
+$LOCA["ru"]["GALATOOL_TOTAL"] = "Всего : #1";
+$LOCA["ru"]["GALATOOL_NOTES"] = "Если у игрока прирост/спад был более 30.000 очков, то он указывается в скобках рядом с именем.";
+$LOCA["ru"]["GALATOOL_FIND"] = "Найти все планеты игрока : ";
+$LOCA["ru"]["GALATOOL_SEARCH"] = "Поиск";
+$LOCA["ru"]["GALATOOL_NOT_UPDATED"] = "Обновления ещё не было";
+
+
+?>
