+++ game/mods/GalaxyTool/manifest.json
@@ -1,0 +1,7 @@
+{
+  "name": "GalaxyTool",
+  "version": "1.0.0",
+  "author": "ogamespec",
+  "description": "Integrated Galaxytool",
+  "website": "https://github.com/ogamespec/ogame-opensource"
+}
