+++ game/mods/GalaxyTool/pages/.htaccess
@@ -1,0 +1,1 @@
+Order allow,deny Deny from all
