+++ game/temp/page_admin.md5
@@ -1,1 +1,1 @@
-a:26:{s:21:"pages_admin/admin.php";s:32:"ea88843401ad4727c6bfccf4241b8812";s:26:"pages_admin/admin_bans.php";s:32:"bf254b309368eb498955a9e27f1a1826";s:28:"pages_admin/admin_battle.php";s:32:"bebe19f795c413bf502ae9cf5168a4a9";s:29:"pages_admin/admin_botedit.php";s:32:"adb6513ab7b7fbd1b62c66b32ce73718";s:26:"pages_admin/admin_bots.php";s:32:"17b8cdc836c463f3f323670f47d931e9";s:31:"pages_admin/admin_broadcast.php";s:32:"b9f2a98d9e22a813aea9f30cd602b87c";s:28:"pages_admin/admin_browse.php";s:32:"9646ea3d78a350800f686a5a84f9ff74";s:30:"pages_admin/admin_checksum.php";s:32:"a5bba6a95a5ca30bd92bc5f1ceafd62a";s:37:"pages_admin/admin_colony_settings.php";s:32:"ee4aac203e4b2f59774cb304c44e192b";s:29:"pages_admin/admin_coupons.php";s:32:"194767ba515d81b25c685526352f2b0e";s:24:"pages_admin/admin_db.php";s:32:"71bd537270edcb492d63aabbc8a8f453";s:27:"pages_admin/admin_debug.php";s:32:"234cc9ad914f82c84667627b845a14d2";s:28:"pages_admin/admin_errors.php";s:32:"4d0eb742d697a5e282fb691116218aed";s:32:"pages_admin/admin_expedition.php";s:32:"47f645bf187f7531e95d6c6b118693c6";s:31:"pages_admin/admin_fleetlogs.php";s:32:"b0836a1576d00b29c9ae9d757240434a";s:26:"pages_admin/admin_loca.php";s:32:"b60cd6eb404e8b1f72033a35ac054024";s:28:"pages_admin/admin_logins.php";s:32:"3ba8e63116a3ecebe470f61e8253f48e";s:29:"pages_admin/admin_planets.php";s:32:"e307c38903429bfb4cb5c66ded513c28";s:27:"pages_admin/admin_queue.php";s:32:"1b1589181d1a4585c5eaa963aaf58975";s:28:"pages_admin/admin_raksim.php";s:32:"48d4c8ab182e2eabb97e1445abf4b1ba";s:29:"pages_admin/admin_reports.php";s:32:"d74853afe468e8d7f5bafc407db6e2cd";s:25:"pages_admin/admin_sim.php";s:32:"27c4b22b4b70e1092a4f20b457c10031";s:25:"pages_admin/admin_uni.php";s:32:"b00d8bc28487594cc541e2b42e21e1fe";s:30:"pages_admin/admin_userlogs.php";s:32:"3276ad33a050751d4b3d130c61c341d8";s:27:"pages_admin/admin_users.php";s:32:"63ed12955b0582815a8d6da0a9af3c66";s:26:"pages_admin/admin_mods.php";s:32:"3f0a4efcf190dbc4dde5f81146156d91";}
+a:26:{s:21:"pages_admin/admin.php";s:32:"ea88843401ad4727c6bfccf4241b8812";s:26:"pages_admin/admin_bans.php";s:32:"3f6de8e52ffbb549e33bf17ff2025583";s:28:"pages_admin/admin_battle.php";s:32:"bebe19f795c413bf502ae9cf5168a4a9";s:29:"pages_admin/admin_botedit.php";s:32:"adb6513ab7b7fbd1b62c66b32ce73718";s:26:"pages_admin/admin_bots.php";s:32:"a64482efd14e8a4f72b7a16eb195a1ca";s:31:"pages_admin/admin_broadcast.php";s:32:"020be97be1e903718df388328d6f3929";s:28:"pages_admin/admin_browse.php";s:32:"9646ea3d78a350800f686a5a84f9ff74";s:30:"pages_admin/admin_checksum.php";s:32:"61b338861735f796523ce7b0c46db6d5";s:37:"pages_admin/admin_colony_settings.php";s:32:"ee4aac203e4b2f59774cb304c44e192b";s:29:"pages_admin/admin_coupons.php";s:32:"a07d9620aa57117393509418bc429962";s:24:"pages_admin/admin_db.php";s:32:"71bd537270edcb492d63aabbc8a8f453";s:27:"pages_admin/admin_debug.php";s:32:"234cc9ad914f82c84667627b845a14d2";s:28:"pages_admin/admin_errors.php";s:32:"4d0eb742d697a5e282fb691116218aed";s:32:"pages_admin/admin_expedition.php";s:32:"47f645bf187f7531e95d6c6b118693c6";s:31:"pages_admin/admin_fleetlogs.php";s:32:"098b479912e691101775132e032d5bc2";s:26:"pages_admin/admin_loca.php";s:32:"b60cd6eb404e8b1f72033a35ac054024";s:28:"pages_admin/admin_logins.php";s:32:"3ba8e63116a3ecebe470f61e8253f48e";s:29:"pages_admin/admin_planets.php";s:32:"fde8fb5aecae7e2fff6813189186d320";s:27:"pages_admin/admin_queue.php";s:32:"f2e87487fa0fb8931203bdfe5a592774";s:28:"pages_admin/admin_raksim.php";s:32:"48d4c8ab182e2eabb97e1445abf4b1ba";s:29:"pages_admin/admin_reports.php";s:32:"d74853afe468e8d7f5bafc407db6e2cd";s:25:"pages_admin/admin_sim.php";s:32:"0b599d9661c982f216f5d2964771c78d";s:25:"pages_admin/admin_uni.php";s:32:"b00d8bc28487594cc541e2b42e21e1fe";s:30:"pages_admin/admin_userlogs.php";s:32:"3276ad33a050751d4b3d130c61c341d8";s:27:"pages_admin/admin_users.php";s:32:"6d7ae172dfc6cd40de5c8e52d62627ff";s:26:"pages_admin/admin_mods.php";s:32:"3f0a4efcf190dbc4dde5f81146156d91";}
