+++ game/temp/reg.md5
@@ -1,1 +1,1 @@
-a:8:{s:26:"reg/check_registration.php";s:32:"c0c20547a3ebd711bcf0034e4383b628";s:17:"reg/errorpage.php";s:32:"e49482e7239e441fcd3d8e98f434c793";s:15:"reg/fa_pass.php";s:32:"af9ee7a23bbf71e326547066da9ca8ef";s:14:"reg/login2.php";s:32:"88216fb1c21ab0a173b354a504f086c4";s:12:"reg/mail.php";s:32:"83d701855dacc29c55945aa3f0ce7c1b";s:11:"reg/new.php";s:32:"ef0a5f5d40b4ccef734d67b1112541a8";s:19:"reg/newredirect.php";s:32:"a1ad562887922a6eeca838f3255d3c2e";s:13:"reg/login.php";s:32:"556e25ee39ba15e15816b860ce95b772";}
+a:8:{s:26:"reg/check_registration.php";s:32:"c0c20547a3ebd711bcf0034e4383b628";s:17:"reg/errorpage.php";s:32:"c938a085ef3d55804ffa74289fc8a80b";s:15:"reg/fa_pass.php";s:32:"ef4dec5ac9af3400eb4507e5c62b57b7";s:14:"reg/login2.php";s:32:"88216fb1c21ab0a173b354a504f086c4";s:12:"reg/mail.php";s:32:"bb090136565c73f138cc33aaf00a6b32";s:11:"reg/new.php";s:32:"725eebc1a523ba9de4d71707dbe9e5d2";s:19:"reg/newredirect.php";s:32:"3ab01c1e159a1c93072d3c58b48776e9";s:13:"reg/login.php";s:32:"556e25ee39ba15e15816b860ce95b772";}
